requirements.txt
python-telegram-bot==20.4
g4f
aiohttp
nest_asyncio
